backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
PyJWT==2.8.0
SQLAlchemy==1.4.52
Werkzeug==2.2.3
